Move StyleSetProps into its own files
