+++ src/StyleSet.cpp
@@ -22,7 +22,6 @@ THE SOFTWARE.
 
 #include "StyleSet.hpp"
 
-#include "Convert.hpp"
 #include "Log.hpp"
 #include "StyleEngine.hpp"
 #include "Warnings.hpp"
@@ -32,7 +31,6 @@ SUPPRESS_WARNINGS
 #include <QtQuick/QQuickItem>
 RESTORE_WARNINGS
 
-#include <iterator>
 #include <string>
 
 namespace aqt
@@ -128,179 +126,7 @@ UiItemPath traversePathUp(QObject* pObj)
   return traverseParentChain<UiItemPath>(pObj, collectPathVisitor);
 }
 
-PropertyMap effectivePropertyMap(const UiItemPath& path, StyleEngine& engine)
-{
-  using std::begin;
-  using std::end;
-  using std::prev;
-
-  auto props = engine.matchPath(path);
-
-  if (path.size() > 1) {
-    const auto ancestorProps =
-      effectivePropertyMap({begin(path), prev(end(path))}, engine);
-
-    if (props.empty()) {
-      props = ancestorProps;
-    } else {
-      props.insert(begin(ancestorProps), end(ancestorProps));
-    }
-  }
-
-  return props;
-}
-
 } // anon namespace
-
-StyleSetProps::StyleSetProps(QObject* pParent)
-  : QObject(pParent)
-{
-}
-
-void StyleSetProps::initStyleSet(const UiItemPath& path, StyleEngine* pEngine)
-{
-  const bool isDiffEngine = mpEngine != pEngine;
-
-  if (isDiffEngine || mPath != path) {
-    if (mpEngine && isDiffEngine) {
-      disconnect(
-        mpEngine, &StyleEngine::styleChanged, this, &StyleSetProps::onStyleChanged);
-    }
-
-    mpEngine = pEngine;
-    mPath = path;
-
-    if (mpEngine && isDiffEngine) {
-      connect(mpEngine, &StyleEngine::styleChanged, this, &StyleSetProps::onStyleChanged);
-    }
-
-    onStyleChanged();
-  }
-}
-
-bool StyleSetProps::isValid() const
-{
-  return !mProperties.empty();
-}
-
-bool StyleSetProps::isSet(const QString& key) const
-{
-  return mProperties.find(key) != mProperties.end();
-}
-
-bool StyleSetProps::getImpl(Property& prop, const QString& key) const
-{
-  PropertyMap::const_iterator it = mProperties.find(key);
-  if (it != mProperties.end()) {
-    prop = it->second;
-    return true;
-  }
-
-  if (!mpEngine.isNull()) {
-    styleSheetsLogWarning() << "Property " << key.toStdString() << " not found ("
-                            << pathToString(mPath) << ")";
-    Q_EMIT mpEngine->exception(QString::fromLatin1("propertyNotFound"),
-                               QString::fromLatin1("Property '%1' not found (%2)")
-                                 .arg(key, QString::fromStdString(pathToString(mPath))));
-  }
-
-  return false;
-}
-
-QVariant StyleSetProps::get(const QString& key) const
-{
-  Property prop;
-  getImpl(prop, key);
-
-  if (prop.mValues.size() == 1) {
-    auto conv = convertProperty<QString>(prop.mValues[0]);
-    if (conv) {
-      return QVariant::fromValue(*conv);
-    }
-  } else if (prop.mValues.size() > 1) {
-    QVariantList result;
-    for (const auto& propValue : prop.mValues) {
-      auto conv = convertProperty<QString>(propValue);
-      if (conv) {
-        result.push_back(conv.get());
-      }
-    }
-
-    return result;
-  }
-
-  return QVariant();
-}
-
-QVariant StyleSetProps::values(const QString& key) const
-{
-  Property prop;
-  getImpl(prop, key);
-
-  if (prop.mValues.size() == 1) {
-    return convertValueToVariant(prop.mValues[0]);
-  }
-
-  return convertValueToVariantList(prop.mValues);
-}
-
-QColor StyleSetProps::color(const QString& key) const
-{
-  return lookupProperty<QColor>(key);
-}
-
-QFont StyleSetProps::font(const QString& key) const
-{
-  return lookupProperty<QFont>(key);
-}
-
-double StyleSetProps::number(const QString& key) const
-{
-  return lookupProperty<double>(key);
-}
-
-bool StyleSetProps::boolean(const QString& key) const
-{
-  return lookupProperty<bool>(key);
-}
-
-QString StyleSetProps::string(const QString& key) const
-{
-  return lookupProperty<QString>(key);
-}
-
-QUrl StyleSetProps::url(const QString& key) const
-{
-  Property prop;
-  auto url = lookupProperty<QUrl>(prop, key);
-
-  if (mpEngine) {
-    auto baseUrl = prop.mSourceLoc.mSourceLayer == 0 ? mpEngine->defaultStyleSheetSource()
-                                                     : mpEngine->styleSheetSource();
-    return mpEngine->resolveResourceUrl(baseUrl, url);
-  }
-
-  return url;
-}
-
-void StyleSetProps::onStyleChanged()
-{
-  loadProperties();
-}
-
-void StyleSetProps::loadProperties()
-{
-  if (mpEngine) {
-    mProperties = effectivePropertyMap(mPath, *mpEngine);
-
-    QObject* pParent = parent();
-    if (qobject_cast<StyleSet*>(pParent) != nullptr) {
-      // emit the change notification from the attached property, otherwise
-      // the QML world won't see it.
-      Q_EMIT qobject_cast<StyleSet*>(parent())->propsChanged();
-    }
-  }
-}
 
 StyleSet::StyleSet(QObject* pParent)
   : QObject(pParent)
